+++ module-services/service-appmgr/data/MmiActionsParams.cpp
@@ -5,6 +5,21 @@
 
 using namespace app::manager::actions;
 
+void MMICustomResultParams::addMessage(const MMIResultMessage &message)
+{
+    messages.push_back(message);
+}
+
+auto MMICustomResultParams::getMessage() -> std::vector<MMIResultMessage>
+{
+    return std::move(messages);
+}
+
+auto MMICustomResultParams::getType() -> MMIType
+{
+    return type;
+}
+
 MMIParams::MMIParams(std::string mmiData) : mmiData{std::move(mmiData)}
 {}
 
@@ -16,7 +31,15 @@ std::string MMIParams::getData() const
 MMIResultParams::MMIResultParams(MMIResult result) : result(result)
 {}
 
+MMIResultParams::MMIResultParams(std::shared_ptr<IMMICustomResultParams> cResult) : customResult(cResult)
+{}
+
 MMIResultParams::MMIResult MMIResultParams::getData() const noexcept
 {
     return result;
+}
+
+auto MMIResultParams::getCustomMessage() const noexcept -> std::shared_ptr<IMMICustomResultParams>
+{
+    return customResult;
 }